api/pyproject.toml
[project]
name = "plant-manager-api"
version = "0.1.0"
description = "Plant watering management system API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "pyjwt>=2.9",
    "bcrypt>=4.1",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "aiomqtt>=2.3",
    "apscheduler>=3.10,<4",
]
